Refactor out internal class backing opaque API objects
Move classes to separate compilation units so the can be used in the 
state object
