+++ keyboardprocessor/src/km_kbp_context_api.cpp
@@ -1,31 +1,12 @@
 #include <cassert>
 #include <algorithm>
 #include <iterator>
-#include <list>
 #include <vector>
 
 #include <keyboardprocessor.h>
 
-#include <utfcodec.hpp>
-
-namespace km {
-namespace kbp
-{
-
-class context : public std::list<km_kbp_context_item>
-{
-  std::vector<value_type> mutable _pres;
-
-public:
-  const_pointer data() const
-  {
-    _pres.assign(begin(), end());
-    return _pres.data();
-  }
-};
-
-}
-}
+#include "context.hpp"
+#include "utfcodec.hpp"
 
 struct km_kbp_context : public km::kbp::context {};
